requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
gunicorn==21.2.0
python-dotenv==1.0.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-WTF==1.1.1
WTForms==3.0.1
email-validator==2.0.0
qrcode[pil]==7.4.2
python-dateutil==2.9.0
pytz==2023.3
openpyxl==3.1.2
psycopg2-binary==2.9.7
Pillow==10.0.1
